+++ app/src/main/java/com/imhungry/jjongseol/ui/meeting/pager/MeetingRecordScreen.kt
@@ -1,35 +1,23 @@
 package com.imhungry.jjongseol.ui.meeting.pager
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
 import com.imhungry.jjongseol.data.chat.ChatMessage
 import com.imhungry.jjongseol.ui.component.ChatBubble
+import com.imhungry.jjongseol.ui.component.CheckItem
 import com.imhungry.jjongseol.ui.component.MeetingTerminationNotification
 import com.imhungry.jjongseol.ui.component.Notification
 import com.imhungry.jjongseol.ui.component.TopSheet
@@ -136,54 +124,5 @@ fun MeetingRecordScreen() {
                 )
             }
         }
-    }
-}
-
-@Composable
-fun CheckItem(
-    text: String,
-    checked: Boolean,
-    isFocused: Boolean,
-    onToggle: () -> Unit
-) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 16.dp)
-    ) {
-        Box(
-            modifier = Modifier.clickable(
-                indication = null,
-                interactionSource = remember { MutableInteractionSource() },
-                onClick = onToggle
-            )
-        ) {
-            Checkbox(
-                checked = checked,
-                onCheckedChange = null,
-                colors = CheckboxDefaults.colors(
-                    checkedColor = Color(0xFFCBE7FD),
-                    uncheckedColor = Color(0xFF2196F3),
-                    checkmarkColor = Color.White,
-                )
-            )
-        }
-
-        Spacer(modifier = Modifier.width(12.dp))
-
-        Text(
-            text = text,
-            style = MaterialTheme.typography.bodyLarge.copy(
-                fontWeight = if (isFocused) FontWeight.Bold else FontWeight.Normal,
-                textDecoration = if (checked) TextDecoration.LineThrough else TextDecoration.None
-            ),
-            color = when {
-                checked -> Color(0xFFB7B7B7)
-                isFocused -> Color.Gray
-                else -> Color(0xFFB7B7B7)
-            },
-            modifier = Modifier.padding(bottom = 2.dp)
-        )
     }
 }
