+++ app/src/main/java/com/imhungry/jjongseol/ui/splash/SplashScreen.kt
@@ -30,9 +30,9 @@ fun SplashScreen(
     LaunchedEffect(Unit) {
         delay(2000)
         if (isLoggedIn == true) {
-            onFinish(SilRokNavigation.Home)
+            onFinish(SilRokNavigation.CompletedMeeting)
         } else {
-            onFinish(SilRokNavigation.Home)
+            onFinish(SilRokNavigation.CompletedMeeting)
         }
     }
